ENGR00217505-7 uboot: gpmi: update the mtd->ecc_stats when reading page failed
If tell the real correcting infomation to the upper layer of
MTD, the torture thread of UBIFS will do the torture test in
a very often frequency. This will eat up all the reservation blocks
of the UBIFS.

So tell the real correcting infomation only when the failure occured,
or the corrected times nearly reached the ECC threshold.

Signed-off-by: Huang Shijie <b32955@freescale.com>